Create README.md
Demeter frontier systems is an expanding consulting company that provides a wide variety of services. Consulting on 

- Construction design and Custom projects 

- Multispectral sensors / Lidar 

- Hydroponics & Vertical farming systems 

- Agriculture drone imaging

 - Procurement Specialist 

 Our primary goal is to contribute to sustainable agriculture by providing self-contained module systems featuring recirculating aquaculture. In addition to our agricultural focus, we aspire to create educational value and grow a variety of fresh greens and microgreens for communities in rural and challenging climates. 

With advancements in machine learning and portable RNA sequencing, we have been able to enhance phenotyping and disease mitigation through careful observation of plant growth. We believe that the thoughtful use of multispectral and hyperspectral sensors can further enhance our understanding of plant life.